move non-constant evaluation in ExprEvaluator dedicated class
