+++ lib/PhpParser/ExprEvaluationException.php
@@ -1,0 +1,6 @@
+<?php declare(strict_types=1);
+
+namespace PhpParser;
+
+class ExprEvaluationException extends \Exception {
+}
